backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pyserial==3.5
python-json-logger==2.0.7
aiofiles==23.2.1
websockets==12.0
pydantic==2.4.2
numpy==1.24.3
